+++ src/main/java/me/despical/commonsbox/number/NumberUtils.java
@@ -30,4 +30,18 @@ public class NumberUtils {
 	public static boolean isBetween(int value, int min, int max) {
 		return value >= min && value <= max;
 	}
+
+	/**
+	 * Serialize int to use it in Inventories size
+	 * ex.you have 38 kits and it will serialize it to 45 (9 * 5)
+	 * because it is valid inventory size
+	 * next ex. you have 55 items and it will serialize it to 63 (9 * 7) not 54 because it's too less
+	 *
+	 * @param i integer to serialize
+	 * @return serialized number
+	 */
+	public static int serializeInt(Integer i) {
+		if (i == 0) return 9;
+		return (i % 9) == 0 ? i : (i + 9 - 1) / 9 * 9;
+	}
 }